requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.3.0
Django==4.2
django-crispy-forms==2.0
django-heroku==0.3.1
django-import-export==3.2.0
django-reset-migrations==0.4.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.6
PyYAML==6.0
reportlab==3.6.12
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.5.0
tzdata==2023.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
